fix : cleanup uploadlink component
