+++ src/app/homepage/links/UserLinks.tsx
@@ -2,13 +2,15 @@
 import React, { useState } from "react";
 import Link from "./Link";
 
-type Props = {};
+type Props = {
+  closeAddLinkSection : () => void;
+};
 
-const UserLinks = (props: Props) => {
+const UserLinks = ({closeAddLinkSection}: Props) => {
   const [link, setLink] = useState<any>(null)
   const [activeLink, setActiveLink] = useState<string>('')
   return <ul>
-    <Link />
+    <Link onCloseAddLink = {closeAddLinkSection} />
   </ul>
 };
